requirements.txt
numpy==2.0.2
pandas==2.2.3
Pillow==11.2.1
scipy==1.13.1
tensorflow==2.19.0
streamlit==1.45.1
matplotlib==3.9.4
